requirements.txt
TavilyClient==1.0.0
requests==2.26.0
python-dotenv==0.19.2
pytest==6.2.5
openai>=1.0.0